chore: move classes in appropriate packages
